ref: return metadata proof to core
